+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_commands.py
@@ -351,13 +351,26 @@ class ContainerappDaprTests(ScenarioTest):
 
         create_containerapp_env(self, env_name, resource_group)
 
-        self.cmd('containerapp create -g {} -n {} --environment {}'.format(resource_group, ca_name, env_name))
+        self.cmd('containerapp create -g {} -n {} --environment {} --dapr-app-id containerapp --dapr-app-port 800 --dapr-app-protocol grpc -dhmrs 4 -dhrbs 50 --dapr-log-level debug --enable-dapr'.format(resource_group, ca_name, env_name), checks=[
+            JMESPathCheck('properties.configuration.dapr.appId', "containerapp"),
+            JMESPathCheck('properties.configuration.dapr.appPort', 800),
+            JMESPathCheck('properties.configuration.dapr.appProtocol', "grpc"),
+            JMESPathCheck('properties.configuration.dapr.enabled', True),
+            JMESPathCheck('properties.configuration.dapr.httpReadBufferSize', 50),
+            JMESPathCheck('properties.configuration.dapr.httpMaxRequestSize', 4),
+            JMESPathCheck('properties.configuration.dapr.logLevel', "debug"),
+            JMESPathCheck('properties.configuration.dapr.enableApiLogging', False),
+        ])
 
-        self.cmd('containerapp dapr enable -g {} -n {} --dapr-app-id containerapp1 --dapr-app-port 80 --dapr-app-protocol http'.format(resource_group, ca_name, env_name), checks=[
+        self.cmd('containerapp dapr enable -g {} -n {} --dapr-app-id containerapp1 --dapr-app-port 80 --dapr-app-protocol http -dal -dhmrs 6 -dhrbs 60 --dapr-log-level warn'.format(resource_group, ca_name, env_name), checks=[
             JMESPathCheck('appId', "containerapp1"),
             JMESPathCheck('appPort', 80),
             JMESPathCheck('appProtocol', "http"),
             JMESPathCheck('enabled', True),
+            JMESPathCheck('httpReadBufferSize', 60),
+            JMESPathCheck('httpMaxRequestSize', 6),
+            JMESPathCheck('logLevel', "warn"),
+            JMESPathCheck('enableApiLogging', True),
         ])
 
         self.cmd('containerapp show -g {} -n {}'.format(resource_group, ca_name), checks=[
@@ -365,6 +378,10 @@ class ContainerappDaprTests(ScenarioTest):
             JMESPathCheck('properties.configuration.dapr.appPort', 80),
             JMESPathCheck('properties.configuration.dapr.appProtocol', "http"),
             JMESPathCheck('properties.configuration.dapr.enabled', True),
+            JMESPathCheck('properties.configuration.dapr.httpReadBufferSize', 60),
+            JMESPathCheck('properties.configuration.dapr.httpMaxRequestSize', 6),
+            JMESPathCheck('properties.configuration.dapr.logLevel', "warn"),
+            JMESPathCheck('properties.configuration.dapr.enableApiLogging', True),
         ])
 
         self.cmd('containerapp dapr disable -g {} -n {}'.format(resource_group, ca_name, env_name), checks=[
@@ -372,6 +389,10 @@ class ContainerappDaprTests(ScenarioTest):
             JMESPathCheck('appPort', 80),
             JMESPathCheck('appProtocol', "http"),
             JMESPathCheck('enabled', False),
+            JMESPathCheck('httpReadBufferSize', 60),
+            JMESPathCheck('httpMaxRequestSize', 6),
+            JMESPathCheck('logLevel', "warn"),
+            JMESPathCheck('enableApiLogging', True),
         ])
 
         self.cmd('containerapp show -g {} -n {}'.format(resource_group, ca_name), checks=[
@@ -379,6 +400,10 @@ class ContainerappDaprTests(ScenarioTest):
             JMESPathCheck('properties.configuration.dapr.appPort', 80),
             JMESPathCheck('properties.configuration.dapr.appProtocol', "http"),
             JMESPathCheck('properties.configuration.dapr.enabled', False),
+            JMESPathCheck('properties.configuration.dapr.httpReadBufferSize', 60),
+            JMESPathCheck('properties.configuration.dapr.httpMaxRequestSize', 6),
+            JMESPathCheck('properties.configuration.dapr.logLevel', "warn"),
+            JMESPathCheck('properties.configuration.dapr.enableApiLogging', True),
         ])